refactor: refactoring professores.js and deleted deleteProfessore.js
