+++ src/controller/professores.js
@@ -1,4 +1,4 @@
-const { buscarTodosProfessores } = require("../model/professores");
+const { buscarTodosProfessores, deletarProfessor } = require("../model/professores");
 
 exports.buscarTodosProfessores = (req, res) => {
   buscarTodosProfessores((error, results) => {
@@ -6,6 +6,18 @@ exports.buscarTodosProfessores = (req, res) => {
       res.status(500).send({ error: "Ocorreu um erro inesperado!" });
     } else {
       res.status(200).send(results.rows);
+    }
+  });
+};
+
+exports.deletarProfessor = (req, res) => {
+  const professorId = req.params.id;
+
+  deletarProfessor(professorId, (error) => {
+    if (error) {
+      res.status(500).send({ error: error });
+    } else {
+      res.status(200).send({ mensagem: "Professor removido com sucesso!" });
     }
   });
 };
