+++ src/model/professores.js
@@ -4,6 +4,13 @@ function buscarTodosProfessores(callback) {
   database.query("SELECT * FROM professores", callback);
 }
 
+function deletarProfessor(id, callback) {
+  const query = "DELETE FROM professores WHERE id = $1;";
+
+  database.query(query, [id], callback);
+}
+
 module.exports = {
   buscarTodosProfessores,
+  deletarProfessor,
 };
